+++ src/Main.java
@@ -4,8 +4,8 @@ public class Main {
         double heightKg = 1.87;
         double weightM = 98;
         BmiService service = new BmiService();
-        double bmi = service.calculate(heightKg, weightM);
-        System.out.println((int) bmi);
+        int bmi = service.calculate(heightKg, weightM);
+        System.out.println(bmi);
 
     }
 }
